reuse test code
